backend/requirements.txt
Django==4.2.21
djangorestframework==3.16.0
django-extensions==4.1
django-filter==25.1
Pillow==11.2.1
shortuuid==1.0.13
python-dotenv==1.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
djoser==2.3.1
dj-database-url==2.3.0




